General converter function for MSstatsPTM.

Requires experimental data, type of spectral processing tool used, and acquisition type. Additionally, the columns containing the protein name and modifications must be indicated

globalConverterPTM(
  data.ptm,
  data.protein = NULL,
  annotation = NULL,
  modification.column,
  processing.tool,
  acquisition.type
)

Arguments

data.ptm
data containing modifications from spectral processing tool

data.protein
data containing unmodified peptides from spectral processing tool

annotation
annotation file if required by MSstats converter. Please view documentation from MSstats or MSstatsTMT for your specific tool.

modification.column
string of column name containing modification information for data.ptm dataset.

processing.tool
name of spectral processing tool for data. Must by one of "SKYLINE", "PROGENESIS", "SPECTRONAUT", "PD", "OPENMS", "OPENSWATH".

acquisition.type
string of "LabelFree" or "TMT". For non-TMT experiments please use "LabelFree"