Annotate modified sites with associated peptides

PTMlocate annotates modified sites with associated peptides.

locatePTM(peptide, uniprot, fasta, modResidue, modSymbol, rmConfound = FALSE)

Arguments

peptide
A string vector of peptide sequences. The peptide sequence does not include its preceding and following AAs.

uniprot
A string vector of Uniprot identifiers of the peptides' originating proteins. UniProtKB entry isoform sequence is used.

fasta
A data.table with FASTA information. Output of tidyFasta.

modResidue
A string. Modifiable amino acid residues.

modSymbol
A string. Symbol of a modified site.

rmConfound
A logical. TRUE removes confounded unmodified sites, FALSE otherwise. Default is FALSE.

Value

A data frame with three columns: uniprot_iso, peptide, site.

Examples

fasta <- tidyFasta("https://www.uniprot.org/uniprot/O13297.fasta")
locatePTM("DRVSYIHNDSC*TR", "O13297", fasta, "C", "\\*")

#> Warning: Both 'col_seq' and '..col_seq' exist in calling scope. Please remove the '..col_seq' variable in calling scope for clarity.

#> Warning: Both 'col_fasta' and '..col_fasta' exist in calling scope. Please remove the '..col_fasta' variable in calling scope for clarity.

#> Warning: Both 'col_res' and '..col_res' exist in calling scope. Please remove the '..col_res' variable in calling scope for clarity.

#>    uniprot_iso        peptide site
#> 1:      O13297 DRVSYIHNDSC*TR C467